Add connection link update on sync consumer.writers
