+++ src/client/error.go
@@ -2,6 +2,7 @@ package client
 
 import (
 	"fmt"
+	"github.com/NubeIO/nubeio-rubix-lib-helpers-go/pkg/rest/v1/rest"
 	"github.com/go-resty/resty/v2"
 )
 
@@ -9,6 +10,19 @@ import (
 type Error struct {
 	Code    int    `json:"error_code,omitempty"`
 	Message string `json:"error_message,omitempty"`
+}
+
+func failedResponse(err error, resp *resty.Response) error {
+	if err != nil {
+		return err
+	}
+	if resp.Error() != nil {
+		return getAPIError(resp)
+	}
+	if rest.StatusCodesAllBad(resp.StatusCode()) {
+		return getAPIError(resp)
+	}
+	return nil
 }
 
 // Convert error response into error message
